Move counter to separate module
